Extract layout query code into query.rs
